consortium-v2: fix the system transaction's sender check in ApplyTransaction (#405)
* consortium-v2: check that sender of system transaction is EOA

Following EIP-3607 ("Reject transactions from senders with deployed code"), if
the sender of transaction has deployed code (not an EOA), the transaction is
rejected. This commit adds the same check to system transaction.

* consortium-v2: cleanup the check of system transaction's sender with coinbase

Currently, the system transaction's sender check looks weird, it still allows
that sender is different from coinbase. However, the sender is checked to be
equal to coinbase in IsSystemMessage already, so there must be no issue here.
This commit fixes that check in ApplyTransaction, we keep this check to make it
more explicit though the check is redundant.