precompiled: add contract for sorting validator with random beacon
This new precompiled contract provides requestSortValidatorSet and
pickValidatorSet function. The requestSortValidatorSet receives the list of all
validator candidates, pick non-rotating validators based on staked amounts and
rotating validators based on weight derived from random beacon. The list of
validators for all epoch of that period is stored in precompiled contract's
storage. The pickValidatorSet supports query the validator set of an epoch in
that period.