consortium-v2: enable pick validator with beacon at Tripp hardfork - 1
Since precompiled contract has no code, it can be marked as empty account if
nonce and balance is 0 too. If the account is empty, the account node is deleted
so writing to storage of that account has no effect. We avoid this by setting
nonce of precompiled contract to 1 at Tripp hardfork block - 1 before the
WrapUpEpoch system transaction.