consortium/v2: Change rules to choose best parent
Current rule to choose best parent block does not take into consideration
the justified block height of the canonical chain and new chain (if
upcoming block has smaller height than that of canonical chain). This
might lead to in-turn miner producing block that cannot be inserted into
canonical chain because it does not satisfy all the conditions for a
reorg.