crypto/bls: disable blst by default (#410)
* crypto/bls: remove unused herumi library

This library is deprecated in prysm and unused in Ronin, so this commit removes
that library.

* crypto/bls: add missing functions to bls stub file

The bls stub helps to make compilation successful when blst is disabled. This
commit adds some missing functions that are used but not included in the stub.

* crypto/bls: disable blst by default

Blst is the library Ronin uses to support BLS signature. The library is
implemented in C and assembly which requires CGO in Go and correct flag to build
the portable version across different architectures.

Other services may use Ronin as a library and don't need to use BLS feature.
However, currently, they still needs CGO and correct CGO flag, which is
cumbersome.

This commit disables blst by default so other services that use Ronin as a
library do not need to care about BLS. Ronin standalone build will explicitly
enable BLS feature.