consortium-v2, precompiled: add contract for sorting validator with random beacon (#419)
* precompiled: add contract for sorting validator with random beacon

This new precompiled contract provides requestSortValidatorSet and
pickValidatorSet function. The requestSortValidatorSet receives the list of all
validator candidates, pick non-rotating validators based on staked amounts and
rotating validators based on weight derived from random beacon. The list of
validators for all epoch of that period is stored in precompiled contract's
storage. The pickValidatorSet supports query the validator set of an epoch in
that period.

* consortium-v2: enable pick validator with beacon at Tripp hardfork - 1

Since precompiled contract has no code, it can be marked as empty account if
nonce and balance is 0 too. If the account is empty, the account node is deleted
so writing to storage of that account has no effect. We avoid this by setting
nonce of precompiled contract to 1 at Tripp hardfork block - 1 before the
WrapUpEpoch system transaction.