consortium-v2: support consensus address change in profile contract (#425)
* consortium-v2: support consensus address change in profile contract

This commit adds a step to query profile id from consensus address to support
the consensus change feature. It also switches the function call to get BLS
public key to another function due to the breaking change in getId2Profile.

* consortium/contract: create contract call helper function

Make the contract call a helper function to clean up the code.

* consortium-v2: pass **big.Int into contract call's output

The maxValidatorNumber returns a *big.Int as an output. In order to get the
output value, we must pass an **big.Int to the contratCall not *big.Int.

Co-authored-by: andicrypt <keepgoing0901@gmail.com>

* consortium-v2: add unit tests for new contract call functions

---------

Co-authored-by: andicrypt <keepgoing0901@gmail.com>