chaincmd: correctly create consortium engine in import chain (#371)
* consortium: remove unused genesis in consensus engine

* chaincmd: correctly create consortium engine in import chain

Currently, when import chain is used, we don't recognize the chain config to
create the consortium engine. This commit creates the consortium engine when it
is set in the chain config.

* consortium-v1: add an option to disable checkpoint header check

In version 1, the checkpoint header check happens early in the header
verification step. This makes statedb read sometimes happens in incorrect
statedb. While full sync, this leads to peer drop but still be able to continue
syncing with other peers. However, while running import chain command or fast
sync, this could lead to error. Add a option to disable checkpoint header check
in these cases.