cmd/flags: update max vote amount per block to 64
After Tripp, the voter set is extended to all validator candidate, whose
upper bound is 64. However, currently, the maximum number of votes per
block is constrained by the flag votepool.maxcurvoteperblock (of value 22),
it should be updated to 64 in default setting. For the same reason, the
maxFutureVoteAmountPerBlock hardcode is also updated from 50 to 64.