eth/filters: return error when getLogs on not found block
Currently, when calling eth_getLogs on not found block, RPC does not return
error but returns empty log. As a result, the user cannot differentiate between
not found block and block with no logs. This commit makes eth_getLogs return an
error when the queried block is not found.

Example of error response after this commit

	{
		"jsonrpc": "2.0",
		"id": 1,
		"error": {
			"code": -32000,
			"message": "block 36077340 not found"
		}
	}