core/vm: move readOnly check outside the TSTORE opcode (#539)
In Ronin, currently we don't have commit 9393d1fb5d ("core/vm: Move
interpreter.ReadOnly check into the opcode implementations"). So the readonly
check is still in the main interpreter loop, not in the opcode implementation.
This commit makes the TSTORE opcode consistent with that behavior in the other
opcodes.