crypto/kzg4844: cache the verify blob proof result
There are 2 places that verify blob proof: transaction pool when a blob
transaction is added, blob sidecar of a new block. The blob in a blob
transaction in pool is very likely to be included in a mined block in the near
future. So we can cache the verification result to avoid double verification.

The benchmark results show significant improvement in cache hit case and
insignificant overhead in cache miss case.

Benchmark with cache hit

                       │    old.txt     │               new.txt                │
                       │     sec/op     │    sec/op     vs base                │
GoKZGVerifyBlobProof-8   2813.49µ ± 15%   19.74µ ± 17%  -99.30% (p=0.000 n=10)

Benchmark with cache miss

                                │   old.txt   │            new.txt            │
                                │   sec/op    │   sec/op     vs base          │
GoKZGVerifyBlobProofCacheMiss-8   4.194m ± 7%   4.391m ± 6%  ~ (p=0.165 n=10)