core: skip gas check for system transaction when tracing
As system transaction has 0 gas price, it does not pass the check that gas price
must be greater or equal to base fee when base fee is not 0. Skip gas check for
system transaction in TransitionDb, system transaction will go through
TransitionDb only in tracing path not in normal transaction processing path.