cmd,eth,docker,rawdb,v1/consortirum: fix missing passing scheme when init genesis and avoid referencing same object when passing parents in cosortium v1 (#608)
* cmd,eth: fix wrong compare logic when data dir is empty and moving checking error correctly

* docker: passing state.scheme when initing the genesis data

* rawdb: add missing freezer in collections

* v1/consortium:  create a copy to keep parents content

In snapshot function, the list parents is popped out gradually for getting its contents, so when calling apply, the parents list is empty. Simply create a copy at the beginning to fix it.
This has been fixed in consortium v2. For a full sync scenario, however, the first blocks are still processed with consortium v1, which causes our node to panic.

---------

Co-authored-by: Francesco4203 <longngocpro@gmail.com>