consortium-v2: make the period definition consistent with contract (#638)
* consortium-v2: make the period definition consistent with contract

Currently, Ronin detects the start of new period is the first block of an epoch
where that block's timestamp is on the different date from first block of
previous epoch. However, in contract, the start of new period is the first block
of epoch where the parent of it (the last block of previous epoch)'s timestamp
is on the different date from the last block of "previous previous" epoch. In
this commit, we change the Ronin logic to be consistent with contract after
Venoki hardfork.

* consortium-v2: remove Tripp hardfork check in IsPeriodBlock

IsPeriodBlock is only called when passing Tripp hardfork already so we don't
need to check it further inside the function.