eth/tracers: make transaction trace response compatible with geth
The current transactionHash field in transaction trace response should be named
txHash to be compatible with go-ethereum. This commit adds txHash instead of
replacing the current transactionHash field and make them have the same value.
This can help to avoid breaking change for current users of these methods since
users are unlikely to use strict DisallowUnknownFields when decoding the json
response.