all: convert time.After to time.NewTimer to avoid leak
In Go before 1.23, we need to call timer.Stop on the timer to allow it to be
collected by garbage collector. So using the pattern like <-time.After(...) can
lead to memory leak. Convert this pattern to time.NewTimer so that we can call
Stop on created timer. This commit does not fix the pattern in tests and unused
code (e.g. light, clique, ...)

Reported-by: eugenioclrc <eugenioclrc@gmail.com>